arch: Move compiler inside devito/arch
