+++ devito/arch/__init__.py
@@ -1,1 +1,2 @@
-from .archinfo import *
+from .archinfo import *  # noqa
+from .compiler import *  # noqa
